src/requirements.txt
PyGithub
python-dotenv
pandas
unidecode
